+++ script/View/SaveButton.ts
@@ -6,16 +6,23 @@ class SaveButton extends DOMView {
 
     constructor(game: Phaser.Game, private constants: CONSTANTS.SaveButton, models: Object) {
         super(game, constants, models);
+        this.setView();
+        this.setEvent();
+    }
+
+    private setView() {
+        this.$.append($(`<img src=${this.constants.image} />`).addClass("buttonImage")
+            .css({ width: "70px", height: "50px" }));
+    }
+
+    private setEvent() {
         this.$.on(this.game.device.touch ? "touchstart" : "mousedown", () => { this.save(); });
-        this.$.css("background-color", "darkcyan");
     }
 
     private save() {
+        if (localStorage.getItem("music") && !confirm("The music you have already saved will be disposed. Is it OK?")) return;
         var str = JSON.stringify(this.music.getMusic);
-        
-        // Local Strage Save
-        localStorage.setItem("music", str);
-
+        localStorage.setItem("music", str); // Local Strage Save
         alert("Your music was saved!");
     }
 }
